requirements.txt
torch
transformers
datasets
numpy
pillow
requests
soundfile
faiss-cpu
gradio
IPython